implemented reworked rito
