+++ examples/new_example.py
@@ -6,12 +6,13 @@ from datetime import datetime
 # Here's a quick example of utilizing the validator and outputting the result to a txt file
 def main():
     validator = two_rito.Validator('r4')
-    resource = open("../test/data/invalid_r4_patient.json", 'r').read()
+    # resource = open("../test/data/invalid_r4_patient.json", 'r').read()
+    # resource = open("../test/data/invalid_r4_patient.json", 'r').read()
 
-    json_resource = json.loads(resource)
-    # output = validator.verbose_validate(json_resource, identifier="invalid_r4_patient.json")
-    output = validator.dir_validate("../test/data/r4", verbose=True)
-    # output = validator.file_validate("../test/data/invalid_r4_patient.json", verbose=True)
+    # json_resource = json.loads(resource)
+    # output = validator.verbose_validate(json_resource)
+    # output = validator.dir_validate("../test/data/r4", verbose=True)
+    output = validator.file_validate("../test/data/invalid_r4_patient.json", verbose=True)
 
     filename = "output" + datetime.now().strftime("%m%d-%H%M-%S") + ".txt"
     with open(filename, "w") as file:
