+++ test/data/invalid_r4_patient.json
@@ -1,5 +1,5 @@
 {
-  "resourceType": "Patient",
+  "resourceType": "Patien",
   "id": "example",
   "text": {
     "status": "generated",
