+++ test/test_validation.py
@@ -21,19 +21,19 @@ def fixture_validator(request):
     return request.param
 
 
-def test_single_resource_valid(validator, resource):
-    fhir_version = validator.get_fhir_version()
+def test_valid_fast_file_validate(validator, resource):
+    fhir_version = validator.fhir_version
     resource_name = f'{resource}.json'
     LOGGER.info(f'Validator Version :: {fhir_version}')
     LOGGER.info(f'Resource File :: {resource}')
 
-    result = validator.fhir_validate(DIRECTORY_PATH + f'/{fhir_version}/{resource}.json')
+    result = validator.file_validate(DIRECTORY_PATH + f'/{fhir_version}/{resource}.json')
     assert result[resource_name] is True
 
 
-def test_directory_validation(validator):
-    fhir_version = validator.get_fhir_version()
+def test_valid_fast_dir_validate(validator):
+    fhir_version = validator.fhir_version
     LOGGER.info(f'Validator Version :: {fhir_version}')
 
-    result = validator.fhir_validate(DIRECTORY_PATH + f'/{fhir_version}')
+    result = validator.dir_validate(DIRECTORY_PATH + f'/{fhir_version}')
     assert result == {'observation.json': True, 'organization.json': True, 'patient.json': True}
